src/service/requirements.txt
djangorestframework==3.7.7
Markdown==2.6.10
djangorestframework-jwt==1.8.0
